+++ src/actions/index.ts
@@ -1,0 +1,3 @@
+import setupAction, {SetupOptions} from './setup'
+
+export {setupAction, SetupOptions}
