+++ src/main.ts
@@ -1,140 +1,15 @@
-/* eslint-disable @typescript-eslint/explicit-function-return-type */
 import * as core from '@actions/core'
-import * as github from '@actions/github'
-import * as os from 'os'
-import * as tc from '@actions/tool-cache'
-import type HeadersInit from 'node-fetch'
-import fetch from 'node-fetch'
-import {promises as fs} from 'fs'
-import retry from 'async-retry'
-
-const getRelease = async (
-  octokit: ReturnType<typeof github.getOctokit>,
-  tag: string
-) => {
-  const owner = 'massdriver-cloud'
-  const repo = 'massdriver-cli'
-  if (tag === 'latest') {
-    return octokit.rest.repos.getLatestRelease({owner, repo})
-  } else {
-    return octokit.rest.repos.getReleaseByTag({
-      owner,
-      repo,
-      tag
-    })
-  }
-}
-
-type GetReleaseResult = ReturnType<typeof getRelease> extends Promise<infer T>
-  ? T
-  : never
-
-type Asset = GetReleaseResult['data']['assets'][0]
-
-interface FetchAssetFileOptions {
-  readonly id: number
-  readonly outputPath: string
-  readonly owner: string
-  readonly repo: string
-  readonly token: string
-}
-
-const baseFetchAssetFile = async (
-  octokit: ReturnType<typeof github.getOctokit>,
-  {id, outputPath, owner, repo, token}: FetchAssetFileOptions
-) => {
-  const {
-    body,
-    headers: {accept, 'user-agent': userAgent},
-    method,
-    url
-  } = octokit.request.endpoint(
-    'GET /repos/:owner/:repo/releases/assets/:asset_id',
-    {
-      asset_id: id,
-      headers: {
-        accept: 'application/octet-stream'
-      },
-      owner,
-      repo
-    }
-  )
-  let headers: HeadersInit = {
-    accept,
-    authorization: `token ${token}`
-  }
-  if (typeof userAgent !== 'undefined')
-    headers = {...headers, 'user-agent': userAgent}
-
-  const response = await fetch(url, {body, headers, method})
-  if (!response.ok) {
-    const text = await response.text()
-    core.warning(text)
-    throw new Error('Invalid response')
-  }
-  const blob = await response.blob()
-  const arrayBuffer = await blob.arrayBuffer()
-
-  await fs.writeFile(outputPath, new Uint8Array(arrayBuffer))
-}
-
-const fetchAssetFile = async (
-  octokit: ReturnType<typeof github.getOctokit>,
-  options: FetchAssetFileOptions
-) =>
-  retry(async () => baseFetchAssetFile(octokit, options), {
-    retries: 5,
-    minTimeout: 1000
-  })
-
-const install = async (target: string): Promise<void> => {
-  const pathToCLI = await tc.extractTar(target)
-  core.addPath(pathToCLI)
-}
-
-const filterByFileName = (file: string) => (asset: Asset) => file === asset.name
-
-const determineArch = (): string => {
-  const arch: string = os.arch()
-  const mappings: {[key: string]: string} = {
-    x64: 'amd64'
-  }
-  return mappings[arch] || arch
-}
-
-const determineFile = (tag: string): string => {
-  // we publish darwin and linux binaries
-  const osPlatform = os.platform()
-  // we publish arm64 and amd64 binaries
-  const osArch = determineArch()
-  return `mass-${tag}-${osPlatform}-${osArch}.tar.gz`
-}
+import {setupAction, SetupOptions} from './actions'
 
 const main = async (): Promise<void> => {
-  const owner = 'massdriver-cloud'
-  const repo = 'massdriver-cli'
   const token = core.getInput('token', {required: false})
-  let tag = core.getInput('tag', {required: false})
+  const tag = core.getInput('tag', {required: false})
 
-  const octokit = github.getOctokit(token)
-  const release = await getRelease(octokit, tag)
-
-  tag = tag === 'latest' ? release.data.tag_name : tag
-  const file = determineFile(tag)
-  const outputPath = `${process.env['RUNNER_TOOL_CACHE']}/${file}`
-  const assetFilterFn = filterByFileName(file)
-  const assets = release.data.assets.filter(assetFilterFn)
-  if (assets.length === 0) throw new Error('Could not find asset id')
-  for (const asset of assets) {
-    await fetchAssetFile(octokit, {
-      id: asset.id,
-      outputPath,
-      owner,
-      repo,
-      token
-    })
+  const opts: SetupOptions = {
+    tag,
+    token
   }
-  install(outputPath)
+  setupAction(opts)
 }
 
 void main()
